# pkg/domain/user.go
package domain

import (
	"time"
)

type User struct {
	ID        string    `json:"id"`
	Name      string    `json:"name"`
	Email     string    `json:"email"`
	Phone     string    `json:"phone"`
	CreatedAt time.Time `json:"created_at"`
	Balance   uint64    `json:"balance"`
}

type RegisterUserRequest struct {
	Name  string `json:"name"`
	Email string `json:"email"`
	Phone string `json:"phone"`
}

type LoginUserRequest struct {
	Name     string `json:"name"`
	Password string `json:"password"`
}